+++ src/messages.rs
@@ -260,6 +260,7 @@ pub struct TimerCustomerMessages {
     pub contact_timeout_leave_message_button: String,
     pub contact_timeout_menu_button: String,
     pub relay_timeout_text: String,
+    pub conversation_inactivity_reset_text: String,
 }
 
 impl ClientMessages {